Added an extra layer to ease the user creation of monitoring windows
